chore: enable line numbers in backtrace
this does not seem to meaningfully impact performance, if it does
it appears to be <1ms. It does increase binary size from 6.3M to 8.4M, but I think the trade-off is worth it for now. With this in place it will be easier
for users to diagnose issues with rtx.