bug: fix "fatal alert BadCertificate"
This makes the version check use http instead of https to avoid
certificate issues. This just has the version number so there isn't a
security risk of a MITM attack here. It also makes the version update
only check once per day in the case of failure or success, before it was
just success.